refactor(vmm): move microvm boot cmdline helpers to mvm-vmm::host
