+++ crates/mvm-runtime/src/driver/spec.rs
@@ -1,6 +1,7 @@
-//! `VmmSpec` is re-exported from `mvm-vmm` for compatibility.
+//! `VmmSpec` re-exported from `mvm-vmm`.
 //!
-//! The backend-agnostic boot recipe now lives in `mvm-vmm::driver::spec`;
-//! `mvm-runtime` re-exports it at the old path for downstream callers.
+//! The backend-agnostic boot recipe lives in `mvm-vmm::driver::spec`;
+//! `mvm-runtime` re-exports it at the old path until downstream callers are
+//! migrated.
 
 pub use mvm_vmm::driver::spec::*;
